requirements.txt
asgiref==3.2.5
boto3==1.17.4
botocore==1.20.4
cachetools==4.2.4
cassandra-driver==3.25.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
CurrencyConverter==0.14.1
defusedxml==0.7.1
dill==0.3.1.1
Django==3.0.4
django-cors-headers==3.2.1
django-nose==1.4.6
# django-spark==0.3.0
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.18.1
et-xmlfile==1.1.0
funcsigs==1.0.2
future==0.16.0
geographiclib==1.50
geomet==0.2.1.post1
geopy==2.1.0
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
graphviz==0.13.2
greenlet==1.1.2
idna==2.10
importlib-metadata==1.6.0
iso8601==0.1.12
itsdangerous==1.1.0
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
nose==1.3.7
numpy==1.19.5
openpyxl==3.0.6
packaging==20.9
pandas==1.1.5
pendulum==1.4.4
prison==0.1.3
protobuf==3.19.1
psutil==5.7.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.6.1
pymongo==3.10.1
pyparsing==3.0.6
pyrsistent==0.16.0
# pyspark==2.4.5
python-daemon==2.1.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.8
s3transfer==0.3.7
six==1.14.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
typing==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
XlsxWriter==3.0.3
zipp==3.1.0
